bugfix: breadcrumbs from last navigation
Update breadcrumbs to leverage the `to` field of NavigationGuard so that
when the BeforeRouteUpdate guard is called the page being navigated to
is able to be referenced so that the breadcrumb field is updated
appropriately.

Without this fix, the breadcrumb will always show the page you navigated
from, and not the current page.

Signed-off-by: Sam Voss <sam.voss@gmail.com>